random: use Lemire-reduction in randomize_page()
randomize_page() is currently slow because it uses a division. It has 4
call sites:

mm/util.c
351:            return randomize_page(mm->brk, SZ_32M);
353:    return randomize_page(mm->brk, SZ_1G);

arch/x86/kernel/sys_x86_64.c
113:                    *begin = randomize_page(*begin, 0x02000000);

arch/x86/kernel/process.c
953:    return randomize_page(mm->brk, 0x02000000);

In all of these, the second argument is smaller than UINT_MAX. That
means the second argument of randomize_page() can be an unsigned int
rather than an unsigned long. This then allows us to make the
randomization of range fast, by using multiplication, as described in
<https://arxiv.org/abs/1805.10941>.

Doing `((u64)a * (u64)b) >> 32` does the same thing as `a % b`,
distribution-wise. It still isn't fair (we'd need rejection sampling for
that), but it is much faster, even on platforms like mips32, where the
`>> 32` results in only a mfhi, without a mflo, or ppc32, where the same
thing happens with a mulhwu with no mullw.

A more appealing approach would have been to observe that all the values
of range are powers of two, and so we can actually just mask off the
value. But not so fast: start might not be page aligned, in which case,
range is adjusted to some stranger value.

Cc: Dominik Brodowski <linux@dominikbrodowski.net>
Cc: Theodore Ts'o <tytso@mit.edu>
Cc: Jason Cooper <jason@lakedaemon.net>
Cc: Kees Cook <keescook@chromium.org>

+++ include/linux/random.h
@@ -110,7 +110,7 @@ declare_get_random_var_wait(int)
 declare_get_random_var_wait(long)
 #undef declare_get_random_var
 
-unsigned long randomize_page(unsigned long start, unsigned long range);
+unsigned long randomize_page(unsigned long start, unsigned int range);
 
 /*
  * This is designed to be standalone for just prandom
